fix : Admin CountUserservice 추상화
